Android | Reuse CurrentImageView.kt for both CameraScreen.kt and BigPictureScreen.kt
