+++ android/src/main/kotlin/com/github/picture2pc/android/ui/main/bigpicturescreen/BigPictureScreen.kt
@@ -2,8 +2,6 @@ package com.github.picture2pc.android.ui.main.bigpicturescreen
 
 import android.graphics.Bitmap
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.detectTransformGestures
 import androidx.compose.foundation.layout.Box
@@ -14,7 +12,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -29,17 +26,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Rect
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.graphics.drawscope.Fill
-import androidx.compose.ui.graphics.drawscope.withTransform
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.onSizeChanged
-import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
 import com.github.picture2pc.android.ui.main.bigpicturescreen.elements.BottomOfScreen
+import com.github.picture2pc.android.ui.main.bigpicturescreen.elements.CurrentImageView
 import com.github.picture2pc.android.ui.util.Settings
 import com.github.picture2pc.android.ui.util.clampInRect
 import com.github.picture2pc.android.viewmodel.camerascreenviewmodels.CameraViewModel
@@ -63,7 +55,7 @@ fun BigPictureScreen(
     var lastScale by remember { mutableFloatStateOf(1f) }
     var lastOffset by remember { mutableStateOf(Offset.Zero) }
     var size by remember { mutableStateOf(Rect(0f, 0f, 0f, 0f)) }
-    var pictureBoxSize by remember { mutableStateOf(IntSize.Zero) }
+
     Box(
         modifier = Modifier
             .fillMaxSize()
@@ -85,7 +77,8 @@ fun BigPictureScreen(
                 )
             }
     ) {
-        Box(
+
+        CurrentImageView(
             modifier = Modifier
                 .align(Alignment.Center)
                 .graphicsLayer(
@@ -94,57 +87,8 @@ fun BigPictureScreen(
                     translationX = offset.x,
                     translationY = offset.y
                 )
-                .clickable(onClick = screenSelectorViewModel::toCamera)
-                .clip(RoundedCornerShape(20.dp))
-                .onSizeChanged { pictureBoxSize = it }
-        ) {
-            if (image != null) {
-                Image(
-                    image.asImageBitmap(),
-                    contentDescription = "Big Picture",
-                    modifier = Modifier.fillMaxSize()
-                )
-            }
-            if (image != null && pictureCorners != null) {
-                Canvas(
-                    modifier = Modifier.fillMaxSize()
-                ) {
-                    val imageAspectRatio = image.width.toFloat() / image.height
-                    val boxAspectRatio = pictureBoxSize.width.toFloat() / pictureBoxSize.height
-
-                    val scaleFactor = if (boxAspectRatio > imageAspectRatio) {
-                        pictureBoxSize.height.toFloat() / image.height
-                    } else {
-                        pictureBoxSize.width.toFloat() / image.width
-                    }
-
-                    val scaledImageSize = Size(
-                        width = image.width * scaleFactor,
-                        height = image.height * scaleFactor
-                    )
-
-                    val topLeftOffset = Offset(
-                        x = (pictureBoxSize.width - scaledImageSize.width) / 2f,
-                        y = (pictureBoxSize.height - scaledImageSize.height) / 2f
-                    )
-                    withTransform({
-                        translate(left = topLeftOffset.x, top = topLeftOffset.y)
-                    }) {
-                        pictureCorners.pointsBox.onEach {
-                            drawCircle(
-                                color = Color.Green,
-                                radius = 10f,
-                                center = Offset(
-                                    (it.x * scaledImageSize.width).toFloat(),
-                                    (it.y * scaledImageSize.height).toFloat()
-                                ),
-                                style = Fill
-                            )
-                        }
-                    }
-                }
-            }
-        }
+                .clickable(onClick = screenSelectorViewModel::toCamera), image, pictureCorners
+        )
 
         Spacer(modifier = Modifier.height(20.dp))
         Column(
