+++ android/src/main/kotlin/com/github/picture2pc/android/ui/main/camerascreen/elements/DisplayImage.kt
@@ -1,7 +1,6 @@
 package com.github.picture2pc.android.ui.main.camerascreen.elements
 
 import android.graphics.Bitmap
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.detectDragGestures
@@ -15,10 +14,11 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
+import com.github.picture2pc.android.data.edgedetection.DetectedBox
+import com.github.picture2pc.android.ui.main.bigpicturescreen.elements.CurrentImageView
 import com.github.picture2pc.android.viewmodel.screenselectorviewmodels.ScreenSelectorViewModel
 import com.github.picture2pc.common.ui.Colors
 import org.koin.compose.rememberKoinInject
@@ -27,14 +27,13 @@ import kotlin.math.roundToInt
 @Composable
 fun DisplayImage(
     image: Bitmap,
+    pictureCorners: DetectedBox?,
     screenSelectorViewModel: ScreenSelectorViewModel = rememberKoinInject()
 ) {
     val imageOffsetX = remember { mutableFloatStateOf(0f) }
     val imageOffsetY = remember { mutableFloatStateOf(0f) }
 
-    Image(
-        bitmap = image.asImageBitmap(),
-        contentDescription = "Taken Picture",
+    CurrentImageView(
         modifier = Modifier
             .getBaseModifier(image.width, image.height)
             .offset {
@@ -58,7 +57,8 @@ fun DisplayImage(
             .clickable(onClick = screenSelectorViewModel::toBigPicture)
             .clip(RoundedCornerShape(20.dp))
             .border(3.dp, color = Colors.PRIMARY, shape = RoundedCornerShape(20.dp))
-            .alpha(0.8f)
+            .alpha(0.8f),
+        image, pictureCorners
     )
 }
